Now sends addMOB for flags

+++ snowman-server/src/main/java/com/sun/darkstar/example/snowman/server/SnowmanPlayer.java
@@ -116,7 +116,7 @@ class SnowmanPlayer implements Serializable, ManagedObject,
                 AppContext.getDataManager().createReference(matcher);
     }
 
-    void setPosition(int x, int y) {
+    void setPosition(float x, float y) {
        this.x = x;
        this.y = y;
     }
